requirements.txt
# astroid==2.3.3
# click==7.1.1
Flask==1.1.2
Flask-MySQLdb==0.2.0
Flask-WTF==0.14.3
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.11.2
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
mysqlclient==1.4.6
passlib==1.7.2
# pylint==2.4.4
python-dotenv==0.13.0
six==1.14.0
typed-ast==1.4.1
Werkzeug==1.0.1
wrapt==1.11.2
WTForms==2.3.1
